# Fix a performance issue when answering "is this tuple Copyable"?