@kavon 2023-05-05
# Summary for main full

No regressions above thresholds

- [Debug-batch](#debug-batch-74a8b7dc26bede6e)
    - [debug-batch brief](#debug-batch-brief-74a8b7dc26bede6e)
    - [debug-batch detailed](#debug-batch-detailed-74a8b7dc26bede6e)
- [Release](#release-74a8b7dc26bede6e)
    - [release brief](#release-brief-74a8b7dc26bede6e)
    - [release detailed](#release-detailed-74a8b7dc26bede6e)



<a name="debug-batch-74a8b7dc26bede6e">

# Debug-batch

<a name="debug-batch-brief-74a8b7dc26bede6e">

## debug-batch brief
<details>
<summary>Regressed (0)</summary>

name | old | new | delta | delta_pct
---: | ---: | ---: | ---: | ---:
</details>
<details>
<summary>Improved (0)</summary>

name | old | new | delta | delta_pct
---: | ---: | ---: | ---: | ---:
</details>
<details>
<summary>Unchanged (delta < 1.0% or delta < 100.0ms) (2)</summary>

name | old | new | delta | delta_pct
---: | ---: | ---: | ---: | ---:
Frontend.NumInstructionsExecuted | 136,862,048,376,595 | 136,774,688,347,110 | -87,360,029,485 | -0.06%
LLVM.NumLLVMBytesOutput | 2,483,803,048 | 2,483,768,296 | -34,752 | -0.0%
</details>

<a name="debug-batch-detailed-74a8b7dc26bede6e">

## debug-batch detailed
<details>
<summary>Regressed (0)</summary>

name | old | new | delta | delta_pct
---: | ---: | ---: | ---: | ---:
</details>
<details>
<summary>Improved (0)</summary>

name | old | new | delta | delta_pct
---: | ---: | ---: | ---: | ---:
</details>
<details>
<summary>Unchanged (delta < 1.0% or delta < 100.0ms) (320)</summary>

name | old | new | delta | delta_pct
---: | ---: | ---: | ---: | ---:
AST.ImportSetCacheHit | 4,968,222 | 4,968,085 | -137 | -0.0%
AST.ImportSetCacheMiss | 793,978 | 793,962 | -16 | -0.0%
AST.ImportSetFoldHit | 402,525 | 402,513 | -12 | -0.0%
AST.ImportSetFoldMiss | 391,453 | 391,449 | -4 | -0.0%
AST.ModuleShadowCacheHit | 56,274 | 55,990 | -284 | -0.5%
AST.ModuleShadowCacheMiss | 4,472 | 4,459 | -13 | -0.29%
AST.ModuleVisibilityCacheHit | 27,649,993 | 27,649,993 | 0 | 0.0%
AST.ModuleVisibilityCacheMiss | 209,659 | 209,659 | 0 | 0.0%
AST.NumASTBytesAllocated | 66,271,305,833 | 66,247,101,162 | -24,204,671 | -0.04%
AST.NumASTScopeExpansions | 8,306,746 | 8,306,764 | 18 | 0.0%
AST.NumASTScopeLookups | 9,531,587 | 9,531,634 | 47 | 0.0%
AST.NumDecls | 340,627 | 340,627 | 0 | 0.0%
AST.NumDependencies | 441,698 | 441,698 | 0 | 0.0%
AST.NumIncrementalDependencies | 25,230 | 25,230 | 0 | 0.0%
AST.NumLinkLibraries | 0 | 0 | 0 | 0.0%
AST.NumLoadedModules | 322,711 | 322,711 | 0 | 0.0%
AST.NumLocalTypeDecls | 6,522 | 6,522 | 0 | 0.0%
AST.NumModuleLookupClassMember | 6,798 | 6,798 | 0 | 0.0%
AST.NumModuleLookupValue | 163,312,359 | 163,307,624 | -4,734 | -0.0%
AST.NumObjCMethods | 93,972 | 93,972 | 0 | 0.0%
AST.NumOperators | 1,028 | 1,028 | 0 | 0.0%
AST.NumPrecedenceGroups | 176 | 176 | 0 | 0.0%
AST.NumReferencedDynamicNames | 159 | 159 | 0 | 0.0%
AST.NumReferencedMemberNames | 14,974,455 | 14,974,453 | -2 | -0.0%
AST.NumReferencedTopLevelNames | 1,342,460 | 1,342,460 | 0 | 0.0%
AST.NumSourceBuffers | 395,204 | 395,204 | 0 | 0.0%
AST.NumSourceLines | 12,354,804 | 12,354,804 | 0 | 0.0%
AST.NumSourceLinesPerSecond | 4,008,461 | 3,989,357 | -19,104 | -0.48%
AST.NumTotalClangImportedEntities | 1,957,377 | 1,956,832 | -545 | -0.03%
Frontend.MaxMallocUsage | 0 | 0 | 0 | 0.0%
Frontend.NumInstructionsExecuted | 136,862,048,376,595 | 136,774,688,347,110 | -87,360,029,485 | -0.06%
Frontend.NumProcessFailures | 31 | 31 | 0 | 0.0%
IRGen.IRGenRequest | 29,478 | 29,478 | 0 | 0.0%
IRGen.OptimizedIRRequest | 0 | 0 | 0 | 0.0%
IRGen.SymbolObjectCodeRequest | 0 | 0 | 0 | 0.0%
IRModule.NumGOTEntries | 297,874 | 297,874 | 0 | 0.0%
IRModule.NumIRAliases | 239,488 | 239,488 | 0 | 0.0%
IRModule.NumIRBasicBlocks | 8,718,749 | 8,718,749 | 0 | 0.0%
IRModule.NumIRComdatSymbols | 0 | 0 | 0 | 0.0%
IRModule.NumIRFunctions | 4,564,520 | 4,564,520 | 0 | 0.0%
IRModule.NumIRGlobals | 4,289,771 | 4,289,771 | 0 | 0.0%
IRModule.NumIRIFuncs | 0 | 0 | 0 | 0.0%
IRModule.NumIRInsts | 108,276,123 | 108,276,123 | 0 | 0.0%
IRModule.NumIRNamedMetaData | 119,432 | 119,432 | 0 | 0.0%
IRModule.NumIRValueSymbols | 8,926,111 | 8,926,111 | 0 | 0.0%
LLVM.NumLLVMBytesOutput | 2,483,803,048 | 2,483,768,296 | -34,752 | -0.0%
Parse.IDEInspectionSecondPassRequest | 0 | 0 | 0 | 0.0%
Parse.NumFunctionsParsed | 396,663 | 396,663 | 0 | 0.0%
Parse.NumIterableDeclContextParsed | 805,230 | 805,231 | 1 | 0.0%
Parse.ParseAbstractFunctionBodyRequest | 684,502 | 684,502 | 0 | 0.0%
Parse.ParseMembersRequest | 617,224 | 617,225 | 1 | 0.0%
Parse.ParseSourceFileRequest | 369,509 | 369,509 | 0 | 0.0%
Parse.ParseTopLevelDeclsRequest | 369,509 | 369,509 | 0 | 0.0%
SILGen.ASTLoweringRequest | 30,813 | 30,813 | 0 | 0.0%
SILGen.ParseSILModuleRequest | 0 | 0 | 0 | 0.0%
SILModule.NumSILGenDefaultWitnessTables | 0 | 0 | 0 | 0.0%
SILModule.NumSILGenFunctions | 3,511,291 | 3,511,291 | 0 | 0.0%
SILModule.NumSILGenGlobalVariables | 140,013 | 140,013 | 0 | 0.0%
SILModule.NumSILGenVtables | 40,379 | 40,379 | 0 | 0.0%
SILModule.NumSILGenWitnessTables | 197,173 | 197,173 | 0 | 0.0%
SILModule.NumSILOptDefaultWitnessTables | 0 | 0 | 0 | 0.0%
SILModule.NumSILOptFunctions | 4,113,532 | 4,113,532 | 0 | 0.0%
SILModule.NumSILOptGlobalVariables | 142,008 | 142,008 | 0 | 0.0%
SILModule.NumSILOptVtables | 40,742 | 40,742 | 0 | 0.0%
SILModule.NumSILOptWitnessTables | 238,740 | 238,740 | 0 | 0.0%
SILOptimizer.ExecuteSILPipelineRequest | 151,423 | 151,423 | 0 | 0.0%
SILOptimizer.LoweredSILRequest | 0 | 0 | 0 | 0.0%
Sema.ABIMembersRequest | 309,251 | 309,251 | 0 | 0.0%
Sema.AbstractGenericSignatureRequest | 145,940 | 145,845 | -95 | -0.07%
Sema.AccessLevelRequest | 19,014,901 | 18,996,096 | -18,805 | -0.1%
Sema.ActorIsolationRequest | 7,119,495 | 7,119,195 | -300 | -0.0%
Sema.AllMembersRequest | 117,887 | 117,887 | 0 | 0.0%
Sema.AnyObjectLookupRequest | 234 | 234 | 0 | 0.0%
Sema.ApplyAccessNoteRequest | 4,365,257 | 4,365,217 | -40 | -0.0%
Sema.AreAllStoredPropertiesDefaultInitableRequest | 49,574 | 49,574 | 0 | 0.0%
Sema.AttachedPropertyWrapperTypeRequest | 30,646 | 30,646 | 0 | 0.0%
Sema.AttachedPropertyWrappersRequest | 23,114,339 | 23,100,128 | -14,210 | -0.06%
Sema.AttachedResultBuilderRequest | 1,880,238 | 1,880,238 | 0 | 0.0%
Sema.BodyInitKindRequest | 94,303 | 94,303 | 0 | 0.0%
Sema.BreakTargetRequest | 6,567 | 6,567 | 0 | 0.0%
Sema.CXXNamespaceMemberLookup | 0 | 0 | 0 | 0.0%
Sema.CallerSideDefaultArgExprRequest | 479,669 | 479,669 | 0 | 0.0%
Sema.CheckDistributedFunctionRequest | 80 | 80 | 0 | 0.0%
Sema.CheckInconsistentAccessLevelOnImport | 59,478 | 59,478 | 0 | 0.0%
Sema.CheckInconsistentImplementationOnlyImportsRequest | 10,759 | 10,759 | 0 | 0.0%
Sema.CheckInconsistentSPIOnlyImportsRequest | 59,478 | 59,478 | 0 | 0.0%
Sema.CheckInconsistentWeakLinkedImportsRequest | 10,759 | 10,759 | 0 | 0.0%
Sema.CheckRedeclarationRequest | 2,331,807 | 2,331,807 | 0 | 0.0%
Sema.ClangCategoryLookupRequest | 0 | 0 | 0 | 0.0%
Sema.ClangDirectLookupRequest | 15,213 | 15,213 | 0 | 0.0%
Sema.ClangRecordMemberLookup | 15,213 | 15,213 | 0 | 0.0%
Sema.ClassAncestryFlagsRequest | 153,075 | 153,063 | -11 | -0.01%
Sema.ClosureEffectsRequest | 152,821 | 152,821 | 0 | 0.0%
Sema.ClosureHasExplicitResultRequest | 58,654 | 58,654 | 0 | 0.0%
Sema.CollectOverriddenDeclsRequest | 15,371,456 | 15,352,229 | -19,226 | -0.13%
Sema.CompareDeclSpecializationRequest | 871,926 | 871,447 | -479 | -0.05%
Sema.CompilerPluginLoadRequest | 0 | 0 | 0 | 0.0%
Sema.ConditionalRequirementsRequest | 879,229 | 879,055 | -174 | -0.02%
Sema.ConformanceHasEffectRequest | 63 | 63 | 0 | 0.0%
Sema.ConstantValueInfoRequest | 0 | 0 | 0 | 0.0%
Sema.ContinueTargetRequest | 2,180 | 2,180 | 0 | 0.0%
Sema.CursorInfoRequest | 0 | 0 | 0 | 0.0%
Sema.CustomAttrNominalRequest | 22,083 | 22,083 | 0 | 0.0%
Sema.CustomAttrTypeRequest | 3,280 | 3,280 | 0 | 0.0%
Sema.CustomRefCountingOperation | 0 | 0 | 0 | 0.0%
Sema.CxxRecordAsSwiftType | 0 | 0 | 0 | 0.0%
Sema.CxxRecordSemantics | 85,795 | 85,785 | -9 | -0.01%
Sema.DefaultAndMaxAccessLevelRequest | 121,636 | 121,636 | 0 | 0.0%
Sema.DefaultArgumentExprRequest | 102,871 | 102,871 | 0 | 0.0%
Sema.DefaultArgumentInitContextRequest | 841 | 841 | 0 | 0.0%
Sema.DefaultArgumentTypeRequest | 21,688 | 21,688 | 0 | 0.0%
Sema.DefaultDefinitionTypeRequest | 8,506 | 8,506 | 0 | 0.0%
Sema.DefaultTypeRequest | 655,399 | 655,399 | 0 | 0.0%
Sema.DerivativeAttrOriginalDeclRequest | 0 | 0 | 0 | 0.0%
Sema.DifferentiableAttributeTypeCheckRequest | 0 | 0 | 0 | 0.0%
Sema.DirectLookupRequest | 81,754,377 | 81,746,117 | -8,260 | -0.01%
Sema.DirectOperatorLookupRequest | 2,578,558 | 2,578,425 | -133 | -0.01%
Sema.DirectPrecedenceGroupLookupRequest | 951,768 | 951,768 | 0 | 0.0%
Sema.DistributedModuleIsAvailableRequest | 67 | 67 | 0 | 0.0%
Sema.DynamicallyReplacedDeclRequest | 7,566,918 | 7,566,618 | -299 | -0.0%
Sema.EnumElementExprPatternRequest | 17 | 17 | 0 | 0.0%
Sema.EnumRawTypeRequest | 41,365 | 41,365 | 0 | 0.0%
Sema.EnumRawValuesRequest | 24,334 | 24,334 | 0 | 0.0%
Sema.ExistentialConformsToSelfRequest | 33,881 | 33,875 | -5 | -0.02%
Sema.ExpandAccessorMacros | 1,302,889 | 1,302,889 | 0 | 0.0%
Sema.ExpandConformanceMacros | 1,298,895 | 1,295,880 | -3,015 | -0.23%
Sema.ExpandMacroExpansionDeclRequest | 0 | 0 | 0 | 0.0%
Sema.ExpandMacroExpansionExprRequest | 0 | 0 | 0 | 0.0%
Sema.ExpandMemberAttributeMacros | 75,289,624 | 75,244,438 | -45,185 | -0.06%
Sema.ExpandPeerMacroRequest | 10,303,663 | 10,303,663 | 0 | 0.0%
Sema.ExpandSynthesizedMemberMacroRequest | 303,176 | 303,176 | 0 | 0.0%
Sema.ExprPatternMatchRequest | 92,535 | 92,535 | 0 | 0.0%
Sema.ExtendedNominalRequest | 706,183 | 706,183 | 0 | 0.0%
Sema.ExtendedTypeRequest | 217,285 | 217,285 | 0 | 0.0%
Sema.ExternalMacroDefinitionRequest | 0 | 0 | 0 | 0.0%
Sema.FragileFunctionKindRequest | 2,790,769 | 2,790,769 | 0 | 0.0%
Sema.FunctionOperatorRequest | 20,338 | 20,338 | 0 | 0.0%
Sema.GenericParamListRequest | 7,403,347 | 7,400,679 | -2,668 | -0.04%
Sema.GenericSignatureRequest | 3,625,353 | 3,625,050 | -303 | -0.01%
Sema.GetDestructorRequest | 47,182 | 47,182 | 0 | 0.0%
Sema.GetDistributedActorArgumentDecodingMethodRequest | 19 | 19 | 0 | 0.0%
Sema.GetDistributedActorIDPropertyRequest | 72 | 72 | 0 | 0.0%
Sema.GetDistributedActorImplicitCodableRequest | 117 | 117 | 0 | 0.0%
Sema.GetDistributedActorInvocationDecoderRequest | 19 | 19 | 0 | 0.0%
Sema.GetDistributedActorSystemPropertyRequest | 63 | 63 | 0 | 0.0%
Sema.GetDistributedActorSystemRemoteCallFunctionRequest | 46 | 46 | 0 | 0.0%
Sema.GetDistributedRemoteCallArgumentInitFunctionRequest | 19 | 19 | 0 | 0.0%
Sema.GetDistributedRemoteCallTargetInitFunctionRequest | 19 | 19 | 0 | 0.0%
Sema.GetDistributedTargetInvocationDecoderDecodeNextArgumentFunctionRequest | 22 | 22 | 0 | 0.0%
Sema.GetDistributedTargetInvocationEncoderRecordArgumentFunctionRequest | 20 | 20 | 0 | 0.0%
Sema.GetDistributedTargetInvocationEncoderRecordErrorTypeFunctionRequest | 11 | 11 | 0 | 0.0%
Sema.GetDistributedTargetInvocationEncoderRecordReturnTypeFunctionRequest | 16 | 16 | 0 | 0.0%
Sema.GetDistributedTargetInvocationResultHandlerOnReturnFunctionRequest | 2 | 2 | 0 | 0.0%
Sema.GetDistributedThunkRequest | 80 | 80 | 0 | 0.0%
Sema.GetImplicitSendableRequest | 264,584 | 263,992 | -592 | -0.22%
Sema.GetRuntimeDiscoverableAttributes | 2,452,984 | 2,452,984 | 0 | 0.0%
Sema.GetSourceFileAsyncNode | 138 | 138 | 0 | 0.0%
Sema.GlobalActorAttributeRequest | 9,042,943 | 9,042,512 | -431 | -0.0%
Sema.GlobalActorInstanceRequest | 1,063 | 1,063 | 0 | 0.0%
Sema.HasCircularInheritedProtocolsRequest | 18,513 | 18,513 | 0 | 0.0%
Sema.HasCircularRawValueRequest | 25,098 | 25,098 | 0 | 0.0%
Sema.HasDefaultInitRequest | 106,862 | 106,862 | 0 | 0.0%
Sema.HasDynamicCallableAttributeRequest | 31,859 | 31,859 | 0 | 0.0%
Sema.HasDynamicMemberLookupAttributeRequest | 581,508 | 581,403 | -105 | -0.02%
Sema.HasImportsMatchingFlagRequest | 733,008 | 733,008 | 0 | 0.0%
Sema.HasIsolatedSelfRequest | 16,027,442 | 16,020,241 | -7,201 | -0.04%
Sema.HasMemberwiseInitRequest | 52,221 | 52,221 | 0 | 0.0%
Sema.HasMissingDesignatedInitializersRequest | 33,330 | 33,330 | 0 | 0.0%
Sema.HasSelfOrAssociatedTypeRequirementsRequest | 16,730 | 16,730 | 0 | 0.0%
Sema.HasUserDefinedDesignatedInitRequest | 106,872 | 106,872 | 0 | 0.0%
Sema.IDEInspectionFileRequest | 0 | 0 | 0 | 0.0%
Sema.InferredGenericSignatureRequest | 242,016 | 242,016 | 0 | 0.0%
Sema.InheritedDeclsReferencedRequest | 7,853,080 | 7,848,386 | -4,694 | -0.06%
Sema.InheritedProtocolsRequest | 866,639 | 866,417 | -222 | -0.03%
Sema.InheritedTypeRequest | 473,483 | 473,387 | -95 | -0.02%
Sema.InheritsSuperclassInitializersRequest | 42,292 | 42,292 | 0 | 0.0%
Sema.InitKindRequest | 197,260 | 197,260 | 0 | 0.0%
Sema.InterfaceTypeRequest | 24,644,448 | 24,628,779 | -15,668 | -0.06%
Sema.IsABICompatibleOverrideRequest | 322,051 | 322,051 | 0 | 0.0%
Sema.IsAccessorTransparentRequest | 576,041 | 576,041 | 0 | 0.0%
Sema.IsActorRequest | 2,525,531 | 2,524,706 | -825 | -0.03%
Sema.IsCallAsFunctionNominalRequest | 6,761 | 6,761 | 0 | 0.0%
Sema.IsDeclApplicableRequest | 0 | 0 | 0 | 0.0%
Sema.IsDeclRefinementOfRequest | 71,755 | 71,755 | 0 | 0.0%
Sema.IsDefaultActorRequest | 63,954 | 63,954 | 0 | 0.0%
Sema.IsDistributedActorRequest | 2,816,123 | 2,815,076 | -1,047 | -0.04%
Sema.IsDynamicRequest | 3,285,740 | 3,285,740 | 0 | 0.0%
Sema.IsFinalRequest | 5,028,298 | 5,025,049 | -3,248 | -0.06%
Sema.IsGetterMutatingRequest | 827,800 | 827,800 | 0 | 0.0%
Sema.IsImplicitlyUnwrappedOptionalRequest | 4,076,079 | 4,076,045 | -34 | -0.0%
Sema.IsMoveOnlyRequest | 784,085 | 784,076 | -8 | -0.0%
Sema.IsNonUserModuleRequest | 104,474 | 104,471 | -3 | -0.0%
Sema.IsObjCRequest | 3,058,426 | 3,058,386 | -40 | -0.0%
Sema.IsSafeUseOfCxxDecl | 0 | 0 | 0 | 0.0%
Sema.IsSetterMutatingRequest | 742,425 | 742,425 | 0 | 0.0%
Sema.IsSingleValueStmtRequest | 374,278 | 374,278 | 0 | 0.0%
Sema.IsStaticRequest | 1,660,833 | 1,660,847 | 14 | 0.0%
Sema.LazyStoragePropertyRequest | 2,573 | 2,573 | 0 | 0.0%
Sema.LocalDiscriminatorsRequest | 551,442 | 551,440 | -2 | -0.0%
Sema.LookupAllConformancesInContextRequest | 3,177,597 | 3,174,189 | -3,408 | -0.11%
Sema.LookupConformanceInModuleRequest | 129,225,720 | 128,848,303 | -377,417 | -0.29%
Sema.LookupInModuleRequest | 11,148,643 | 11,144,137 | -4,505 | -0.04%
Sema.LookupInfixOperatorRequest | 82,324 | 82,324 | 0 | 0.0%
Sema.LookupPostfixOperatorRequest | 74 | 74 | 0 | 0.0%
Sema.LookupPrecedenceGroupRequest | 33,550 | 33,550 | 0 | 0.0%
Sema.LookupPrefixOperatorRequest | 452 | 452 | 0 | 0.0%
Sema.MacroDefinitionRequest | 0 | 0 | 0 | 0.0%
Sema.MangleLocalTypeDeclRequest | 3,149 | 3,149 | 0 | 0.0%
Sema.ModuleImplicitImportsRequest | 10,759 | 10,759 | 0 | 0.0%
Sema.ModuleLibraryLevelRequest | 22 | 22 | 0 | 0.0%
Sema.ModuleQualifiedLookupRequest | 4,058,395 | 4,053,842 | -4,553 | -0.11%
Sema.NamedLazyMemberLoadSuccessCount | 37,929,833 | 37,925,563 | -4,269 | -0.01%
Sema.NamingPatternRequest | 242,247 | 242,247 | 0 | 0.0%
Sema.NeedsNewVTableEntryRequest | 1,072,204 | 1,072,204 | 0 | 0.0%
Sema.NumAccessorBodiesSynthesized | 480,199 | 480,199 | 0 | 0.0%
Sema.NumAccessorsSynthesized | 579,269 | 579,269 | 0 | 0.0%
Sema.NumConformancePathsRecorded | 998,823 | 998,796 | -27 | -0.0%
Sema.NumConformancesDeserialized | 6,268,228 | 6,264,554 | -3,674 | -0.06%
Sema.NumConstraintScopes | 80,318,741 | 80,317,318 | -1,423 | -0.0%
Sema.NumConstraintsConsideredForEdgeContraction | 2,726,085 | 2,726,085 | 0 | 0.0%
Sema.NumCrossImportsChecked | 0 | 0 | 0 | 0.0%
Sema.NumCrossImportsFound | 0 | 0 | 0 | 0.0%
Sema.NumCyclicOneWayComponentsCollapsed | 0 | 0 | 0 | 0.0%
Sema.NumDeclsDeserialized | 120,400,561 | 120,343,763 | -56,798 | -0.05%
Sema.NumDeclsTypechecked | 3,314,099 | 3,314,099 | 0 | 0.0%
Sema.NumLazyIterableDeclContexts | 15,610,815 | 15,607,277 | -3,538 | -0.02%
Sema.NumLazyRequirementSignatures | 1,192,685 | 1,192,611 | -73 | -0.01%
Sema.NumLazyRequirementSignaturesLoaded | 792,818 | 792,736 | -82 | -0.01%
Sema.NumLeafScopes | 61,824,571 | 61,823,536 | -1,035 | -0.0%
Sema.NumRequirementMachineCompletionSteps | 30,061,360 | 30,048,579 | -12,781 | -0.04%
Sema.NumRequirementMachineUnifiedConcreteTerms | 1,372,707 | 1,371,697 | -1,010 | -0.07%
Sema.NumRequirementMachines | 2,354,507 | 2,353,393 | -1,114 | -0.05%
Sema.NumTypesDeserialized | 33,226,833 | 33,209,460 | -17,373 | -0.05%
Sema.NumUnloadedLazyIterableDeclContexts | 8,796,436 | 8,796,333 | -102 | -0.0%
Sema.ObjCInterfaceAndImplementationRequest | 948,400 | 948,401 | 1 | 0.0%
Sema.OpaqueReadOwnershipRequest | 502,884 | 502,884 | 0 | 0.0%
Sema.OpaqueResultTypeRequest | 513 | 513 | 0 | 0.0%
Sema.OperatorPrecedenceGroupRequest | 1,027 | 1,027 | 0 | 0.0%
Sema.OverriddenDeclsRequest | 6,673,966 | 6,671,067 | -2,899 | -0.04%
Sema.ParamSpecifierRequest | 2,243,692 | 2,243,692 | 0 | 0.0%
Sema.PatternBindingEntryRequest | 817,934 | 817,934 | 0 | 0.0%
Sema.PatternTypeRequest | 969,965 | 969,965 | 0 | 0.0%
Sema.PolymorphicEffectKindRequest | 1,120,913 | 1,120,913 | 0 | 0.0%
Sema.PolymorphicEffectRequirementsRequest | 75 | 75 | 0 | 0.0%
Sema.PreCheckResultBuilderRequest | 951 | 951 | 0 | 0.0%
Sema.PreCheckReturnStmtRequest | 595,192 | 595,192 | 0 | 0.0%
Sema.PrimaryAssociatedTypesRequest | 7,876 | 7,876 | 0 | 0.0%
Sema.PrimarySourceFilesRequest | 10,759 | 10,759 | 0 | 0.0%
Sema.PropertyWrapperAuxiliaryVariablesRequest | 2,451,660 | 2,451,658 | -2 | -0.0%
Sema.PropertyWrapperBackingPropertyTypeRequest | 30,646 | 30,646 | 0 | 0.0%
Sema.PropertyWrapperInitializerInfoRequest | 715,935 | 715,935 | 0 | 0.0%
Sema.PropertyWrapperLValuenessRequest | 2,636 | 2,636 | 0 | 0.0%
Sema.PropertyWrapperMutabilityRequest | 821,669 | 821,669 | 0 | 0.0%
Sema.PropertyWrapperTypeInfoRequest | 599 | 599 | 0 | 0.0%
Sema.ProtocolDependenciesRequest | 679,121 | 678,946 | -175 | -0.03%
Sema.ProtocolRequirementsRequest | 81,341 | 81,341 | 0 | 0.0%
Sema.ProtocolRequiresClassRequest | 47,957 | 47,829 | -127 | -0.27%
Sema.ProvideDefaultImplForRequest | 15,371,456 | 15,352,229 | -19,226 | -0.13%
Sema.QualifiedLookupRequest | 10,044,052 | 10,042,623 | -1,429 | -0.01%
Sema.RangeInfoRequest | 0 | 0 | 0 | 0.0%
Sema.RawCommentRequest | 1,184,962 | 1,184,962 | 0 | 0.0%
Sema.RenamedDeclRequest | 5,221 | 5,221 | 0 | 0.0%
Sema.RequirementRequest | 194,117 | 194,117 | 0 | 0.0%
Sema.RequirementSignatureRequest | 852,310 | 852,135 | -175 | -0.02%
Sema.RequiresOpaqueAccessorsRequest | 2,099,833 | 2,099,833 | 0 | 0.0%
Sema.RequiresOpaqueModifyCoroutineRequest | 508,305 | 508,305 | 0 | 0.0%
Sema.ResolveEffectiveMemberwiseInitRequest | 0 | 0 | 0 | 0.0%
Sema.ResolveImplicitMemberRequest | 949,032 | 948,959 | -73 | -0.01%
Sema.ResolveMacroRequest | 24,989 | 24,989 | 0 | 0.0%
Sema.ResolveProtocolNameRequest | 0 | 0 | 0 | 0.0%
Sema.ResolveTypeEraserTypeRequest | 0 | 0 | 0 | 0.0%
Sema.ResolveTypeRequest | 5,758,634 | 5,758,634 | 0 | 0.0%
Sema.ResultBuilderTypeRequest | 555,045 | 555,045 | 0 | 0.0%
Sema.ResultTypeRequest | 1,259,123 | 1,259,123 | 0 | 0.0%
Sema.RootAndResultTypeOfKeypathDynamicMemberRequest | 0 | 0 | 0 | 0.0%
Sema.RootTypeOfKeypathDynamicMemberRequest | 0 | 0 | 0 | 0.0%
Sema.SPIGroupsRequest | 8,810,644 | 8,808,063 | -2,581 | -0.03%
Sema.ScopedImportLookupRequest | 936 | 936 | 0 | 0.0%
Sema.SelfAccessKindRequest | 1,044,396 | 1,044,401 | 5 | 0.0%
Sema.SelfBoundsFromGenericSignatureRequest | 92,500 | 92,500 | 0 | 0.0%
Sema.SelfBoundsFromWhereClauseRequest | 2,042,720 | 2,041,515 | -1,205 | -0.06%
Sema.SemanticAvailableRangeAttrRequest | 8,073 | 8,073 | 0 | 0.0%
Sema.SemanticBriefCommentRequest | 199,396 | 199,396 | 0 | 0.0%
Sema.SemanticUnavailableAttrRequest | 2,962,855 | 2,962,855 | 0 | 0.0%
Sema.SetterAccessLevelRequest | 286,062 | 286,062 | 0 | 0.0%
…